requirements.txt
Bottleneck==1.3.2
certifi==2021.10.8
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-Cors==3.0.10
gunicorn==20.1.0
importlib-metadata==4.10.0
itsdangerous==2.0.1
Jinja2==3.0.3
joblib
MarkupSafe==2.0.1
mkl-fft==1.3.1
mkl-random
mkl-service==2.4.0
numexpr
numpy
pandas
pymongo==4.0
python-dateutil
pytz==2021.3
PyYAML==6.0
scikit-learn
scipy
six
threadpoolctl
typing_extensions==4.0.1
Werkzeug==2.0.2
wincertstore==0.2
zipp==3.6.0
